requirements.txt
chatterbox-tts>=0.1.0
flask>=2.0.0 